requirements.txt
pyserial
scapy
requests
paho-mqtt
pandas
numpy
matplotlib
seaborn
scipy
folium
pyshark
psycopg2
